refactor: user 서비스 레이어 분리
