+++ src/main/java/org/websoso/WSSServer/application/AuthApplication.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.user.service;
+package org.websoso.WSSServer.application;
 
 import static org.websoso.WSSServer.exception.error.CustomAuthError.INVALID_TOKEN;
 
@@ -9,19 +9,28 @@ import org.websoso.WSSServer.config.jwt.CustomAuthenticationToken;
 import org.websoso.WSSServer.config.jwt.JWTUtil;
 import org.websoso.WSSServer.config.jwt.JwtProvider;
 import org.websoso.WSSServer.config.jwt.JwtValidationType;
-import org.websoso.WSSServer.user.domain.RefreshToken;
+import org.websoso.WSSServer.dto.auth.LogoutRequest;
 import org.websoso.WSSServer.dto.auth.ReissueResponse;
+import org.websoso.WSSServer.dto.user.LoginResponse;
 import org.websoso.WSSServer.exception.exception.CustomAuthException;
+import org.websoso.WSSServer.oauth2.service.KakaoService;
 import org.websoso.WSSServer.repository.RefreshTokenRepository;
+import org.websoso.WSSServer.user.domain.RefreshToken;
+import org.websoso.WSSServer.user.domain.User;
+import org.websoso.WSSServer.user.repository.UserDeviceRepository;
+import org.websoso.WSSServer.user.service.UserService;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class AuthService {
-
+public class AuthApplication {
     private final JwtProvider jwtProvider;
     private final JWTUtil jwtUtil;
     private final RefreshTokenRepository refreshTokenRepository;
+    private final UserDeviceRepository userDeviceRepository;
+    private final UserService userService;
+    private final KakaoService kakaoService;
+    private static final String KAKAO_PREFIX = "kakao";
+    private static final String APPLE_PREFIX = "apple";
 
     public ReissueResponse reissue(String refreshToken) {
         RefreshToken storedRefreshToken = refreshTokenRepository.findByRefreshToken(refreshToken)
@@ -40,5 +49,28 @@ public class AuthService {
         refreshTokenRepository.save(new RefreshToken(newRefreshToken, userId));
 
         return ReissueResponse.of(newAccessToken, newRefreshToken);
+    }
+
+    // TODO: getUserOrException -> existUserOrException 변경
+    @Transactional(readOnly = true)
+    public LoginResponse login(Long userId) {
+        User user = userService.getUserOrException(userId);
+
+        CustomAuthenticationToken customAuthenticationToken = new CustomAuthenticationToken(user.getUserId(), null,
+                null);
+        String token = jwtProvider.generateAccessToken(customAuthenticationToken);
+
+        return LoginResponse.of(token);
+    }
+
+    public void logout(User user, LogoutRequest request) {
+        refreshTokenRepository.findByRefreshToken(request.refreshToken())
+                .ifPresent(refreshTokenRepository::delete);
+
+        userDeviceRepository.deleteByUserAndDeviceIdentifier(user, request.deviceIdentifier());
+
+        if (user.getSocialId().startsWith(KAKAO_PREFIX)) {
+            kakaoService.kakaoLogout(user);
+        }
     }
 }
