+++ src/main/java/org/websoso/WSSServer/controller/AuthController.java
@@ -21,23 +21,23 @@ import org.websoso.WSSServer.dto.auth.ReissueResponse;
 import org.websoso.WSSServer.dto.user.WithdrawalRequest;
 import org.websoso.WSSServer.oauth2.service.AppleService;
 import org.websoso.WSSServer.oauth2.service.KakaoService;
-import org.websoso.WSSServer.user.service.AuthService;
-import org.websoso.WSSServer.user.service.UserService;
+import org.websoso.WSSServer.application.AccountApplication;
+import org.websoso.WSSServer.application.AuthApplication;
 
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthApplication authApplication;
     private final KakaoService kakaoService;
     private final AppleService appleService;
-    private final UserService userService;
+    private final AccountApplication accountApplication;
 
     @PostMapping("/reissue")
     public ResponseEntity<ReissueResponse> reissue(@RequestBody ReissueRequest reissueRequest) {
         return ResponseEntity
                 .status(OK)
-                .body(authService.reissue(reissueRequest.refreshToken()));
+                .body(authApplication.reissue(reissueRequest.refreshToken()));
     }
 
     @PostMapping("/auth/login/kakao")
@@ -58,7 +58,7 @@ public class AuthController {
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Void> logout(@AuthenticationPrincipal User user,
                                        @Valid @RequestBody LogoutRequest request) {
-        userService.logout(user, request);
+        authApplication.logout(user, request);
         return ResponseEntity
                 .status(NO_CONTENT)
                 .build();
@@ -68,7 +68,7 @@ public class AuthController {
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Void> withdrawUser(@AuthenticationPrincipal User user,
                                              @Valid @RequestBody WithdrawalRequest withdrawalRequest) {
-        userService.withdrawUser(user, withdrawalRequest);
+        accountApplication.withdrawUser(user, withdrawalRequest);
         return ResponseEntity
                 .status(NO_CONTENT)
                 .build();
